Revert "[ENGG-5292] SOAP Importer (#4356)"
This reverts commit 60f9fb521d2523b6d0c9bed9c3f2f2e9832dde17.
