+++ app/src/features/apiClient/screens/apiClient/components/CommonApiClientImporter/CommonApiClientImporter.tsx
@@ -1,10 +1,11 @@
 import * as Sentry from "@sentry/react";
 import React, { useCallback, useState } from "react";
+import { FilePicker } from "components/common/FilePicker";
 import { HiOutlineExternalLink } from "@react-icons/all-files/hi/HiOutlineExternalLink";
 import { MdInfoOutline } from "@react-icons/all-files/md/MdInfoOutline";
 import { Col, Tooltip } from "antd";
 import { RQAPI, ApiClientImporterType, EnvironmentData } from "@requestly/shared/types/entities/apiClient";
-import { ApiClientImporterMethod } from "@requestly/alternative-importers";
+import { ApiClientImporterMethod, ApiClientImporterOutput } from "@requestly/alternative-importers";
 import { toast } from "utils/Toast";
 import {
   apiRecordsActions,
@@ -12,15 +13,25 @@ import {
   getApiClientFeatureContext,
   useApiClientRepository,
 } from "features/apiClient/slices";
-import { trackImportFailed, trackImportSuccess } from "modules/analytics/events/features/apiClient";
+import { useApiClientDispatch } from "features/apiClient/slices/hooks/base.hooks";
+import {
+  trackImportFailed,
+  trackImportParsed,
+  trackImportParseFailed,
+  trackImportSuccess,
+} from "modules/analytics/events/features/apiClient";
 import "./commonApiClientImporter.scss";
-import { SuccessfulParseView, SuccessfulParseViewProps } from "./components/SuccessfulParseView";
+import { ImportErrorView } from "./components/ImporterErrorView";
+import { SuccessfulParseView } from "./components/SuccessfulParseView";
 import { wrapWithCustomSpan } from "utils/sentry";
 import { SPAN_STATUS_ERROR, SPAN_STATUS_OK } from "@sentry/core";
-import { ImportInputView, ImportFile, LinkViewConfig } from "./components/ImportInputView";
 import { apiRecordsRankingManager } from "features/apiClient/helpers/RankingManager";
 
-export type { ImportFile, LinkViewConfig };
+export interface ImportFile {
+  content: string;
+  name: string;
+  type: string;
+}
 
 export interface CommonApiClientImporterProps {
   productName: string;
@@ -29,9 +40,6 @@ export interface CommonApiClientImporterProps {
   importerType: ApiClientImporterType;
   onImportSuccess: () => void;
   docsLink?: string;
-  linkView?: LinkViewConfig;
-  renderSuccessView?: (props: SuccessfulParseViewProps) => React.ReactNode;
-  renderLoadingView?: () => React.ReactNode;
 }
 
 export const CommonApiClientImporter: React.FC<CommonApiClientImporterProps> = ({
@@ -41,23 +49,94 @@ export const CommonApiClientImporter: React.FC<CommonApiClientImporterProps> = (
   importerType,
   onImportSuccess,
   docsLink,
-  linkView,
-  renderSuccessView,
-  renderLoadingView,
 }) => {
   const [isLoading, setIsLoading] = useState<boolean>(false);
+  const [isDataProcessing, setIsDataProcessing] = useState<boolean>(false);
   const [isParseComplete, setIsParseComplete] = useState<boolean>(false);
   const [collectionsData, setCollectionsData] = useState<RQAPI.CollectionRecord[]>([]);
   const [environmentsData, setEnvironmentsData] = useState<EnvironmentData[]>([]);
+  const [importError, setImportError] = useState<string | null>(null);
 
   const { environmentVariablesRepository, apiClientRecordsRepository } = useApiClientRepository();
   const { dispatch } = getApiClientFeatureContext().store;
 
   const handleResetImport = () => {
+    setImportError(null);
     setIsParseComplete(false);
     setCollectionsData([]);
     setEnvironmentsData([]);
   };
+
+  const handleFileDrop = useCallback(
+    async (files: File[]) => {
+      return wrapWithCustomSpan(
+        {
+          name: `[Transaction] api_client.${importerType.toLowerCase()}_import.process_files`,
+          op: `api_client.${importerType.toLowerCase()}_import.process_files`,
+          forceTransaction: true,
+          attributes: {},
+        },
+        async (files: File[]) => {
+          setIsDataProcessing(true);
+          const processFiles = files.map((file) => {
+            return new Promise<ApiClientImporterOutput>((resolve, reject) => {
+              const reader = new FileReader();
+              reader.onerror = () => {
+                reject(new Error("Could not process the selected files! Try again."));
+              };
+              reader.onload = () => {
+                const fileContent = reader.result as string;
+                importer({ content: fileContent, name: file.name, type: file.type })
+                  .then((output) => resolve(output))
+                  .catch((error) => reject(error));
+              };
+              reader.readAsText(file);
+            });
+          });
+
+          await Promise.allSettled(processFiles)
+            .then((results: PromiseSettledResult<ApiClientImporterOutput>[]) => {
+              const hasAllFilesFailed = !results.some((result) => result.status === "fulfilled");
+              if (hasAllFilesFailed) {
+                throw new Error("Could not process the files! Please check if files are valid");
+              }
+              const processedResults: { collections: RQAPI.CollectionRecord[]; environments: EnvironmentData[] } = {
+                collections: [],
+                environments: [],
+              };
+              results.forEach((result) => {
+                if (result.status === "fulfilled") {
+                  processedResults.collections.push(result.value.data.collection);
+                  processedResults.environments.push(...result.value.data.environments);
+                }
+              });
+              if (processedResults.collections.length === 0 && processedResults.environments.length === 0) {
+                throw new Error("Selected Files don't contain any collections or environments");
+              }
+              setCollectionsData(processedResults.collections);
+              setEnvironmentsData(processedResults.environments);
+              setIsParseComplete(true);
+              trackImportParsed(importerType, processedResults.collections.length, null);
+              Sentry.getActiveSpan()?.setStatus({
+                code: SPAN_STATUS_OK,
+              });
+            })
+            .catch((error) => {
+              setImportError(error.message || "Could not process the selected files! Try again.");
+              trackImportParseFailed(importerType, error.message);
+              Sentry.captureException(error);
+              Sentry.getActiveSpan()?.setStatus({
+                code: SPAN_STATUS_ERROR,
+              });
+            })
+            .finally(() => {
+              setIsDataProcessing(false);
+            });
+        }
+      )(files);
+    },
+    [importer, importerType]
+  );
 
   const handleImportEnvironments = useCallback(
     async (environments: EnvironmentData[]) => {
@@ -305,110 +384,100 @@ export const CommonApiClientImporter: React.FC<CommonApiClientImporterProps> = (
     [importRootCollection]
   );
 
-  const handleImportData = useCallback(
-    async (collectionsOverride?: RQAPI.CollectionRecord[], environmentsOverride?: EnvironmentData[]) => {
-      return wrapWithCustomSpan(
-        {
-          name: `[Transaction] api_client.${importerType.toLowerCase()}_import.import_data`,
-          op: `api_client.${importerType.toLowerCase()}_import.import_data`,
-          forceTransaction: true,
-          attributes: {},
-        },
-        async () => {
-          setIsLoading(true);
-          try {
-            const collectionsToImport = collectionsOverride || collectionsData;
-            const environmentsToImport = environmentsOverride || environmentsData;
+  const handleImportData = useCallback(async () => {
+    return wrapWithCustomSpan(
+      {
+        name: `[Transaction] api_client.${importerType.toLowerCase()}_import.import_data`,
+        op: `api_client.${importerType.toLowerCase()}_import.import_data`,
+        forceTransaction: true,
+        attributes: {},
+      },
+      async () => {
+        setIsLoading(true);
+        try {
+          const importPromises = [];
+          importPromises.push(handleImportEnvironments(environmentsData));
+          importPromises.push(handleImportCollections(collectionsData));
+          const importResults = await Promise.allSettled(importPromises);
 
-            const importPromises = [];
-            importPromises.push(handleImportEnvironments(environmentsToImport));
-            importPromises.push(handleImportCollections(collectionsToImport));
-            const importResults = await Promise.allSettled(importPromises);
+          const environmentResults = importResults[0];
+          const collectionResults = importResults[1];
 
-            const environmentResults = importResults[0];
-            const collectionResults = importResults[1];
+          let totalImported = 0;
+          let totalFailed = 0;
 
-            let totalImported = 0;
-            let totalFailed = 0;
+          if (environmentResults && environmentResults.status === "fulfilled") {
+            const envResults = environmentResults.value as PromiseSettledResult<{ id: string; name: string }>[];
+            const successfulEnvironments = envResults.filter((result) => result.status === "fulfilled").length;
+            totalImported += successfulEnvironments;
+            totalFailed += envResults.length - successfulEnvironments;
+          } else {
+            totalFailed += environmentsData.length;
+          }
 
-            if (environmentResults && environmentResults.status === "fulfilled") {
-              const envResults = environmentResults.value as PromiseSettledResult<{ id: string; name: string }>[];
-              const successfulEnvironments = envResults.filter((result) => result.status === "fulfilled").length;
-              totalImported += successfulEnvironments;
-              totalFailed += envResults.length - successfulEnvironments;
-            } else {
-              totalFailed += environmentsToImport.length;
-            }
+          if (collectionResults && collectionResults.status === "fulfilled") {
+            const collectionResult = collectionResults.value as {
+              success: boolean;
+              importedCount: number;
+              failedCount: number;
+            };
+            totalImported += collectionResult.importedCount;
+            totalFailed += collectionResult.failedCount;
+          } else {
+            totalFailed += collectionsData.length;
+          }
 
-            if (collectionResults && collectionResults.status === "fulfilled") {
-              const collectionResult = collectionResults.value as {
-                success: boolean;
-                importedCount: number;
-                failedCount: number;
-              };
-              totalImported += collectionResult.importedCount;
-              totalFailed += collectionResult.failedCount;
-            } else {
-              totalFailed += collectionsToImport.length;
-            }
-
-            if (totalImported === 0) {
-              Sentry.captureException(new Error("Failed to import collections and environments"));
-              Sentry.getActiveSpan()?.setStatus({
-                code: SPAN_STATUS_ERROR,
-              });
-              return;
-            }
-
-            if (totalFailed === 0) {
-              toast.success(
-                `Successfully imported ${collectionsToImport.length} ${
-                  collectionsToImport.length !== 1 ? "collections" : "collection"
-                } ${
-                  environmentsToImport.length > 0
-                    ? `and ${environmentsToImport.length} ${
-                        environmentsToImport.length !== 1 ? "environments" : "environment"
-                      }`
-                    : ""
-                }`
-              );
-              trackImportSuccess(importerType, totalImported, null);
-              onImportSuccess();
-              Sentry.getActiveSpan()?.setStatus({
-                code: SPAN_STATUS_OK,
-              });
-            } else {
-              toast.warn(
-                `Partially imported: ${totalImported} succeeded, ${totalFailed} failed. Please review your files.`
-              );
-              trackImportSuccess(importerType, totalImported, null);
-              onImportSuccess();
-              Sentry.captureException(new Error("Partially imported collections and environments"));
-              Sentry.getActiveSpan()?.setStatus({
-                code: SPAN_STATUS_ERROR,
-              });
-            }
-          } catch (e) {
-            trackImportFailed(importerType, e.message);
-            Sentry.captureException(e);
+          if (totalImported === 0) {
+            setImportError("Failed to import collections and environments");
+            Sentry.captureException(new Error("Failed to import collections and environments"));
             Sentry.getActiveSpan()?.setStatus({
               code: SPAN_STATUS_ERROR,
             });
-          } finally {
-            setIsLoading(false);
+            return;
           }
+
+          if (totalFailed === 0) {
+            toast.success(
+              `Successfully imported ${collectionsData.length} ${
+                collectionsData.length !== 1 ? "collections" : "collection"
+              } and ${environmentsData.length} ${environmentsData.length !== 1 ? "environments" : "environment"}`
+            );
+            trackImportSuccess(importerType, collectionsData.length, null);
+            onImportSuccess();
+            Sentry.getActiveSpan()?.setStatus({
+              code: SPAN_STATUS_OK,
+            });
+          } else {
+            toast.warn(
+              `Partially imported: ${totalImported} succeeded, ${totalFailed} failed. Please review your files.`
+            );
+            trackImportSuccess(importerType, totalImported, null);
+            onImportSuccess();
+            Sentry.captureException(new Error("Partially imported collections and environments"));
+            Sentry.getActiveSpan()?.setStatus({
+              code: SPAN_STATUS_ERROR,
+            });
+          }
+        } catch (e) {
+          setImportError("Failed to import collections and environments");
+          trackImportFailed(importerType, e.message);
+          Sentry.captureException(e);
+          Sentry.getActiveSpan()?.setStatus({
+            code: SPAN_STATUS_ERROR,
+          });
+        } finally {
+          setIsLoading(false);
         }
-      )();
-    },
-    [
-      collectionsData,
-      environmentsData,
-      handleImportEnvironments,
-      handleImportCollections,
-      onImportSuccess,
-      importerType,
-    ]
-  );
+      }
+    )();
+  }, [
+    collectionsData,
+    environmentsData,
+    handleImportEnvironments,
+    handleImportCollections,
+    onImportSuccess,
+    importerType,
+  ]);
 
   const HeaderComponent: React.FC<{}> = () => {
     return (
@@ -444,47 +513,32 @@ export const CommonApiClientImporter: React.FC<CommonApiClientImporterProps> = (
     );
   };
 
-  const handleDataProcessed = useCallback(
-    (data: { collections: RQAPI.CollectionRecord[]; environments: EnvironmentData[] }) => {
-      setCollectionsData(data.collections);
-      setEnvironmentsData(data.environments);
-      setIsParseComplete(true);
-    },
-    []
-  );
-
   return (
     <div className={`common-api-client-importer-container`}>
       <HeaderComponent />
       <div className="common-api-client-importer-body">
         {isParseComplete ? (
-          renderSuccessView ? (
-            renderSuccessView({
-              collectionsData,
-              environmentsData,
-              isLoading,
-              handleImportData,
-              onBack: handleResetImport,
-            })
-          ) : (
-            <SuccessfulParseView
-              collectionsData={collectionsData}
-              environmentsData={environmentsData}
-              isLoading={isLoading}
-              handleImportData={handleImportData}
-              onBack={handleResetImport}
-            />
-          )
-        ) : (
-          <ImportInputView
-            productName={productName}
-            supportedFileTypes={supportedFileTypes}
-            linkView={linkView}
-            importer={importer}
-            importerType={importerType}
-            onDataProcessed={handleDataProcessed}
-            renderLoadingView={renderLoadingView}
+          <SuccessfulParseView
+            collectionsData={collectionsData}
+            environmentsData={environmentsData}
+            isLoading={isLoading}
+            handleImportData={handleImportData}
           />
+        ) : (
+          <>
+            {importError ? <ImportErrorView importError={importError} /> : null}
+            <FilePicker
+              maxFiles={5}
+              onFilesDrop={(files) => {
+                handleResetImport();
+                handleFileDrop(files);
+              }}
+              isProcessing={isDataProcessing}
+              title={`Drag and drop your ${productName} export file to upload`}
+              subtitle={`Accepted file formats: ${supportedFileTypes.join(", ")}`}
+              selectorButtonTitle={isParseComplete || importError ? "Try another file" : "Select file"}
+            />
+          </>
         )}
       </div>
       <FooterComponent />
