--- app/src/features/apiClient/screens/apiClient/components/modals/SoapImporterModal/soapSuccessfulParseView.scss
@@ -1,83 +1,0 @@
-.soap-success-view {
-  .soap-collection-name {
-    margin-bottom: 16px;
-    color: var(--requestly-color-text-subtle);
-    font-family: var(--font-family-default, Inter);
-    font-size: var(--font-size-sm, 12px);
-    font-style: normal;
-    font-weight: 400;
-    line-height: 18px;
-  }
-
-  .soap-options-container {
-    margin-bottom: 16px;
-    display: flex;
-    flex-direction: column;
-    gap: 12px;
-
-    .soap-option-item {
-      display: flex;
-      flex-direction: column;
-      align-items: flex-start;
-      justify-content: space-between;
-
-      .select-heading {
-        justify-content: left;
-        font-size: 12px;
-        margin-bottom: 4px;
-        display: flex;
-        align-items: center;
-
-        .info-tooltip-icon {
-          margin-left: 6px;
-          color: var(--requestly-color-text-subtle);
-          font-size: 16px;
-          width: 14px;
-          height: 14px;
-        }
-      }
-
-      .ant-select {
-        width: 100%;
-        height: 32px;
-      }
-    }
-  }
-
-  .imported-entities-container {
-    .collection-organisation-info {
-      color: var(--requestly-color-text-subtle);
-      font-size: 11px;
-      margin-left: 8px;
-    }
-  }
-
-  .importer-actions-row {
-    display: flex;
-    justify-content: flex-end;
-    align-items: center;
-    gap: 8px;
-
-    .need-help-link {
-      margin-right: auto;
-      color: var(--requestly-color-text-subtle);
-      font-size: var(--font-size-sm, 12px);
-      text-decoration: none;
-      display: flex;
-      align-items: center;
-      gap: 4px;
-
-      .help-icon {
-        font-size: 16px;
-        width: 14px;
-        height: 14px;
-      }
-
-      &:hover,
-      &:focus {
-        text-decoration: underline;
-        color: var(--requestly-color-text);
-      }
-    }
-  }
-}
