+++ app/src/features/apiClient/screens/apiClient/utils.ts
@@ -446,12 +446,6 @@ export const parseCurlRequest = (curl: string): RQAPI.Request => {
           contentType = RequestContentType.FORM;
         } else if (httpContentType.includes("application/json")) {
           contentType = RequestContentType.JSON;
-        } else if (
-          httpContentType.includes("application/xml") ||
-          httpContentType.includes("text/xml") ||
-          httpContentType.includes("application/soap")
-        ) {
-          contentType = RequestContentType.XML;
         } else {
           contentType = RequestContentType.RAW;
         }
