+++ shared/src/types/entities/apiClient/api.ts
@@ -256,7 +256,6 @@ export enum ApiClientImporterType {
   BRUNO = "BRUNO",
   CURL = "CURL",
   OPENAPI = "OPENAPI",
-  SOAP = "SOAP",
 }
 
 export type CollectionVariableMap = Record<string, { variables: EnvironmentVariables }>;
